+++ src/nanoarrow/nanoarrow_config.h.in
@@ -15,8 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#ifndef NANOARROW_BUILD_ID_H_INCLUDED
-#define NANOARROW_BUILD_ID_H_INCLUDED
+#ifndef NANOARROW_CONFIG_H_INCLUDED
+#define NANOARROW_CONFIG_H_INCLUDED
 
 #define NANOARROW_VERSION_MAJOR @NANOARROW_VERSION_MAJOR@
 #define NANOARROW_VERSION_MINOR @NANOARROW_VERSION_MINOR@
@@ -28,5 +28,8 @@
    NANOARROW_VERSION_PATCH)
 
 @NANOARROW_NAMESPACE_DEFINE@
+
+#define NANOARROW_CXX_NAMESPACE_BEGIN namespace nanoarrow {
+#define NANOARROW_CXX_NAMESPACE_END }
 
 #endif
